New request handling for Node.js & Browser JavaScript; issue #72
This is a big change that has been needed for awhile. The request library superagent has been scrapped in favor of a native in-home solution that will allow for more customization on how the requests are made in Node, and in the Browser. This update has significantly reduced code bloat, and speeds up the calls to the Reddit API. More importantly, Snoocore is now more browser friendly. More tests and configuration are needed in the future, but this is a step in the right direction