major overhaul; more info in desc
removed cookie based login; issue #91

support application only oauth; issue #61

Application only oauth / the removal of cookie based oauth needed to be done in one step. There would be too much noise in the code if we supported both at once. The default behavior is to use Application Only OAuth when making "unauthenticated" calls, and when we need to support authenticated calls, we can call `.auth` as usual. Calling `.deauth` or using `bypassAuth` will switch back to the Application Only OAuth... Snoocore handles both access_tokens behind the scenes renewing them as necessary.

Note that this is in its own branch for now, and may be for some time as it is worked on. Changes in wip/master will be implemented in here as necessary - as of now these are two separate code bases.