Replace endpoint tree with a more minimal endpoint properties.
This saves a huge number of bytes. But comes with a structural overhaul. Before the endpoint tree was massive and was used to *enforce* endpoint rules. This has proven to be a pain to maintain when reddit changes how the endpoints are used, so why not remove it all together?

We now just keep a record of unique properties of endpoints (e.g. if it requires api_type="json"), and assume that the user knows best for the rest including method calls, etc. Helpful errors are still in place to help guide the user however.