remove endpoint "properties"
Before there was a need for endpoint properties when we were parsing paths / had an internal tree. In 3.x.x we allow for a more free form path arrangement, and it is a bit too much effort to try and enforce special properties on endpoints. This may need to be revisited in the future, but for now adding api_type="json" to the arguments of every endpoint keeps things nice without too much overhead.